requirements.txt
langchain
openai
lark-parser
pydantic
python-dotenv
google-generativeai
